+++ CHANGELOG.md
@@ -4,6 +4,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 # Upcoming
 
 ### ✅ Added
+- Use `AppSettings.fileUploadConfig` for setting supported UTI types for the file picker [#713](https://github.com/GetStream/stream-chat-swiftui/pull/713)
 - Colors and images for voice recording view [#704](https://github.com/GetStream/stream-chat-swiftui/pull/704)
   - `ColorPalette.voiceMessageControlBackground`
   - `Images.pauseFilled`